+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - Validation & inline fix utility for scenes that are either not in build or disabled in build.
+- New settings entries for scene in build validation & inline fix utilities.
 
 ### Changed